Created useAppUpdates hook and updated UpdatableApp
